fix: update configurations and resolve CI build issues
