+++ backend/src/app/middleware/auth.middleware.ts
@@ -99,29 +99,19 @@ const auth =
         }
 
         // Verify JWT token
-        const verifiedUser = JwtHelpers.verifyToken(
+        const decodedUser = JwtHelpers.verifyToken(
           token,
           config.jwt.secret as Secret
-        ) as JwtVerifiedUser;
+        );
 
-        if (!verifiedUser?._id) {
+        if (!isJwtVerifiedUser(decodedUser)) {
           throw new ApiError(
             httpStatus.UNAUTHORIZED,
             "Invalid token"
-
-          const decodedUser = JwtHelpers.verifyToken(
-            token,
-            config.jwt.secret as Secret
           );
-
-          if (!isJwtVerifiedUser(decodedUser)) {
-            throw new ApiError(
-              httpStatus.UNAUTHORIZED,
-              "Invalid token"
-            );
         }
 
-      const verifiedUser = decodedUser;
+        const verifiedUser = decodedUser;
 
         const user = await User.findById(verifiedUser._id);
 
@@ -179,11 +169,7 @@ const auth =
       } catch (err) {
         next(err);
       }
-      req.user = user;
-      return next();
-    } catch (err) {
-      return next(err);
-    }
+
   };
 
 export default auth;
