+++ backend/src/app/modules/user/user.model.ts
@@ -104,7 +104,6 @@ export const UserSchema: Schema<IUser> = new Schema<IUser, UserModel>(
 
 UserSchema.pre("save", async function (next) {
   const user = this;
-main
   if (!user.isModified("password")) {
     return next();
   }
@@ -112,12 +111,10 @@ main
     this.passwordChangedAt = new Date(Date.now() - 1000);
   }
 
-main
-    user.password = await bcrypt.hash(
-      user.password,
-      Number(config.bcrypt_salt_rounds)
-    );
-  }
+  user.password = await bcrypt.hash(
+    user.password,
+    Number(config.bcrypt_salt_rounds)
+  );
 
   next();
 });
