+++ backend/src/app/utils/stripEmojis.ts
@@ -1,15 +1,4 @@
 /**
- * Strips all emojis from a string using comprehensive Unicode ranges.
- */
-export const stripEmojis = (input: string): string => {
-  if (input === null || input === undefined) return '';
-  if (typeof input !== 'string') return '';
-  
-  // Regex pattern covering standard emojis, presentation selectors, flags, ZWJ sequences, dingbats, etc.
-  const emojiRegex = /[\u{1F300}-\u{1F9FF}]|[\u{1F600}-\u{1F64F}]|[\u{1F680}-\u{1F6FF}]|[\u{2600}-\u{26FF}]|[\u{2700}-\u{27BF}]|[\u{1F1E6}-\u{1F1FF}]|[\u{1F191}-\u{1F251}]|[\u{1F900}-\u{1F9FF}]|[\u{1FA70}-\u{1FAFF}]|[\u{2b50}]|[\u{1F004}]|[\u{231a}]|[\u{231b}]|[\u{23e9}-\u{23ec}]|[\u{23f0}]|[\u{23f3}]|[\u{25fe}]|[\u{fe0f}]|[\u{200d}]/gu;
-  
-  return input.replace(emojiRegex, '').trim();
-
  * Removes all emoji characters from a string using Unicode emoji ranges.
  * Useful for sanitizing usernames, titles, and content fields where
  * emojis are not desired or could cause display issues.
@@ -19,7 +8,6 @@ export const stripEmojis = (input: string): string => {
  * - Dingbats (2700-27BF)
  * - Transport/map symbols (1F680-1F6FF)
  * - Miscellaneous symbols (1F300-1F5FF)
- * - Emoticons extended (1F600-1F64F)
  * - Symbols and pictographs (1F300-1F9FF)
  * - Flags (1F1E6-1F1FF)
  * - Zero-width joiners used in emoji sequences
@@ -33,6 +21,5 @@ export const stripEmojis = (input: string): string => {
   // eslint-disable-next-line no-control-regex
   const emojiRegex = /[\u{1F600}-\u{1F64F}]|[\u{1F300}-\u{1F5FF}]|[\u{1F680}-\u{1F6FF}]|[\u{1F1E0}-\u{1F1FF}]|[\u{2600}-\u{26FF}]|[\u{2700}-\u{27BF}]|[\u{1F900}-\u{1F9FF}]|[\u{1FA00}-\u{1FA6F}]|[\u{1FA70}-\u{1FAFF}]|[\u{231A}-\u{231B}]|[\u{23E9}-\u{23F3}]|[\u{23F8}-\u{23FA}]|[\u{25AA}-\u{25AB}]|[\u{25B6}]|[\u{25C0}]|[\u{25FB}-\u{25FE}]|[\u{2614}-\u{2615}]|[\u{2648}-\u{2653}]|[\u{267F}]|[\u{2693}]|[\u{26A1}]|[\u{26AA}-\u{26AB}]|[\u{26BD}-\u{26BE}]|[\u{26C4}-\u{26C5}]|[\u{26CE}]|[\u{26D4}]|[\u{26EA}]|[\u{26F2}-\u{26F3}]|[\u{26F5}]|[\u{26FA}]|[\u{26FD}]|[\u{2702}]|[\u{2705}]|[\u{2708}-\u{270D}]|[\u{270F}]|[\u{2712}]|[\u{2714}]|[\u{2716}]|[\u{271D}]|[\u{2721}]|[\u{2728}]|[\u{2733}-\u{2734}]|[\u{2744}]|[\u{2747}]|[\u{274C}]|[\u{274E}]|[\u{2753}-\u{2755}]|[\u{2757}]|[\u{2763}-\u{2764}]|[\u{2795}-\u{2797}]|[\u{27A1}]|[\u{27B0}]|[\u{27BF}]|[\u{2934}-\u{2935}]|[\u{2B05}-\u{2B07}]|[\u{2B1B}-\u{2B1C}]|[\u{2B50}]|[\u{2B55}]|[\u{3030}]|[\u{303D}]|[\u{3297}]|[\u{3299}]|[\u{FE0F}]|[\u{200D}]|[\u{20E3}]|[\u{E0020}-\u{E007F}]/gu;
 
-  return input.replace(emojiRegex, "");
-
+  return input.replace(emojiRegex, "").trim();
 };
