+++ frontend/tsconfig.json
@@ -6,6 +6,6 @@
   ],
   "compilerOptions": {
     "noUnusedLocals": false,
-    "noUnusedParameters": false
+    "noImplicitAny": false,
   }
 }
